Add navigation bar at top of the webpage
Navigation  bar is diveded into 3 parts: The logo of the page, the links
to  navigate  through  the  page and the contact icons that redirect the
user to my social medias.

The  logo  was  hardcoded in style.css , using the 'Black Ops One' font,
imported from google-fonts. It also has an white solid border, with some
padding  and a background using linear-gradient with shades of gray. The
final result is really cool, better than any image i found.

The  links  hover effect  are  really simple, they just half the opacity
within 0.5s