Add project sections
The  main  portifolio page will display 4 projects, they're organized by
using  CSS Grid. Each project tile have a background image corresponding
to  the project. They also have a hover effect that displays the project
name,  technologies used in the projects and 2 buttons that redirects to
that project code and to the live page.